refactor(routines): move the timezone onto the routines list as a gray card (HOU-470)
The timezone is account-wide, so it belongs on the routines list rather than
inside one routine's editor. Render it there as a `bg-secondary` card (matching
the editor's section cards) with a labeled white-well `<select>` + the "All your
routines run in this timezone" hint — directly under the "New routine" row. The
routine editor no longer carries a timezone control; its "next run" preview
still shows the resolved zone. Labels + locale keys move editor -> grid.
