+++ app/src/locales/en/routines.json
@@ -5,7 +5,9 @@
     "emptyTitle": "Set it and forget it",
     "emptyDescription": "Routines fire on a schedule and only ping you when something actually needs attention.",
     "newRoutine": "New routine",
-    "descriptionShort": "Recurring tasks that fire on schedule and only ping you when something needs attention."
+    "descriptionShort": "Recurring tasks that fire on schedule and only ping you when something needs attention.",
+    "timezoneLabel": "Timezone",
+    "timezoneHint": "All your routines run in this timezone."
   },
   "editor": {
     "back": "Back to routines",
@@ -29,8 +31,6 @@
     "sectionWhen": "When it runs",
     "sectionBehavior": "Behavior",
     "sectionRecent": "Recent runs",
-    "timezoneLabel": "Timezone",
-    "timezoneHint": "All your routines run in this timezone.",
     "nextRun": "Next run {relative}",
     "schedulePreview": "Schedule preview",
     "schedulePreviewHint": "Pick a valid schedule to see when this routine will fire.",
